Create a new Shoe Detail destination that includes:
A new layout
A TextView label and EditView for the Shoe Name, Company, Shoe Size and Description
A Cancel button with an action to navigate back to the shoe list screen
A Save button with an action to navigate back to the shoe list screen and add a new Shoe to the Shoe View Model.

Make sure you can’t go back to onboarding screens In the Shoe List screen:
use an Activity level ViewModel to hold a list of Shoes (use by activityViewModels)
Observe the shoes variable from the ViewModel
Add a new layout item into the scrollview for each shoe.